# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  store.PaymentMethod:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact:
    email: mochammad.yogaprasetya112@gmail.com
    name: API Support
    url: http://www.myogaprasetya.my.id
  description: |-
    API ini menggunakan mekanisme rate limiter untuk membatasi jumlah permintaan yang dapat dilakukan oleh setiap klien dalam jangka waktu tertentu.
    Hal ini bertujuan untuk mencegah penyalahgunaan API, memastikan kinerja yang stabil, dan melindungi server dari potensi serangan seperti brute force atau DDoS.
    Rate limiter ini didukung oleh Redis sebagai penyimpanan sementara yang cepat dan andal untuk melacak jumlah permintaan setiap klien berdasarkan API key atau alamat IP mereka.
    Redis digunakan karena kemampuannya yang tinggi dalam menangani data secara real-time dengan latensi rendah.
    Jika batas permintaan terlampaui, klien akan menerima respons dengan kode status 429 (Too Many Requests).
  termsOfService: http://www.myogaprasetya.my.id
  title: Tokopedia API
paths:
...
  /payment-methods:
    get:
      consumes:
      - application/json
      description: Get all available payment methods
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.PaymentMethod'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: List payment methods
      tags:
      - payment-method
    post:
      consumes:
      - application/json
      description: Create a new payment method
      parameters:
      - description: Payment Method
        in: body
        name: payment_method
        required: true
        schema:
          $ref: '#/definitions/store.PaymentMethod'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.PaymentMethod'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create payment method
      tags:
      - payment-method
  /payment-methods/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment method by ID
      parameters:
      - description: Payment Method ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete payment method
      tags:
      - payment-method
    get:
      consumes:
      - application/json
      description: Get payment method by ID
      parameters:
      - description: Payment Method ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.PaymentMethod'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get payment method
      tags:
      - payment-method
    put:
      consumes:
      - application/json
      description: Update an existing payment method
      parameters:
      - description: Payment Method ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Method
        in: body
        name: payment_method
        required: true
        schema:
          $ref: '#/definitions/store.PaymentMethod'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.PaymentMethod'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update payment method
      tags:
      - payment-method
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"